# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BlogFlow API",
        "description": "API REST de la plateforme BlogFlow. Authentification via token Bearer (Laravel Sanctum). Obtenez votre token via POST /api/login.",
        "contact": {
            "name": "Équipe BlogFlow",
            "email": "admin@blogflow.fr"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://blog-api-service-fbnq.onrender.com",
            "description": "Serveur de production"
        },
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/comments": {
            "post": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Poster un commentaire",
                "description": "Ajoute un commentaire sur un article existant. Nécessite d'être authentifié.",
                "operationId": "300619a0a4003920aafe960e45688102",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "visitor_name",
                                    "visitor_email",
                                    "message",
                                    "article_id"
                                ],
                                "properties": {
                                    "visitor_name": {
                                        "description": "Nom du visiteur (max 191 caractères)",
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "visitor_email": {
                                        "description": "Email du visiteur (max 191 caractères)",
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean@exemple.fr"
                                    },
                                    "message": {
                                        "description": "Contenu du commentaire (max 1000 caractères)",
                                        "type": "string",
                                        "example": "Excellent article, très bien expliqué !"
                                    },
                                    "article_id": {
                                        "description": "ID de l'article auquel ce commentaire est rattaché",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Commentaire publié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Commentaire ajouté avec succès !"
                                        },
                                        "comment": {
                                            "$ref": "#/components/schemas/Comment"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/comments/{id}": {
            "delete": {
                "tags": [
                    "Commentaires"
                ],
                "summary": "Supprimer un commentaire",
                "description": "Supprime un commentaire. Seul le propriétaire du commentaire ou un administrateur peut effectuer cette action.",
                "operationId": "721e5dc0c32ef81276203df965767606",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant du commentaire",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Commentaire supprimé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Commentaire supprimé avec succès !"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "This action is unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Commentaire introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Comment": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "visitor_name": {
                        "type": "string",
                        "example": "Jean Dupont"
                    },
                    "visitor_email": {
                        "type": "string",
                        "format": "email",
                        "example": "jean@exemple.fr"
                    },
                    "message": {
                        "type": "string",
                        "example": "Excellent article !"
                    },
                    "article_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The title field is required."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}